+++ RELEASE_NOTES.md
@@ -1,5 +1,15 @@
 # Release Notes
 
+## 8.1.1 - 05/06/2026
+
+### Added
+
+- Added Custom Settings autocomplete and navigation with a `CustomSettings` command prefix.
+- Added Custom Metadata autocomplete and navigation with a `CustomMetadata` command prefix.
+- Added pinning for command history so favorite commands stay above recent commands.
+- Updated Logo
+- Improved UI and button layouts and clicks
+
 ## 8.0.1 - 04/30/2026
 
 ### Fixed